PrimeCachingStatement is newly introduced as new mapped statement as a child of CachingStatement to perform cache warming using prime cache pattern.
User can enable primedCache attribute flag of cacheModel element in
sql-map configuration to make use of prime cache for loading data into
cache.

PrimeCachingStatement holds the reference to mapped statement that
needs to be executed to load data for primed cache. When user set
primedCache attribute to true, it is mandatory to set
primedCacheQuery attribute of select element that want to take
advantage of primed cache. The primedCacheQuery attribute specifies
which query needs to be run to warm the cache.

Primed cache uses partial keys and specific keys. A partial key
corresponds to a discrete set of data with common characteristics
whereas a specific key corresponds to an exact data item in the cache.
The mapped statement of primedCacheQuery is used as partial key.
The mapped statement of select query and parameter of select query
combined together is used as specific key. The keyProperty attribute
will be provided in select element by user to say which property of the
results should be used as parameter component of specific key.

The loadPrimedCache is newly added to
SqlMapClient interface to provide flexibility for user to start cache
warming at any point of time. The SqlMapClient transfer the loading of
primed cache task to SqlMapExecutorDelegate. It maintains the list of
mapped statements. When loadPrimedCache is invoked, it iterates all
mapped statements and finds list of all prime caching statements
defined in sql map configuration. It makes queryForList method for all
primed cache queries to warm the cache with all primed cache data.